update indexes location
